BTS-7422　対処（イベント登録プライオリティ操作）

+++ prog/src/battle/btl_event_factor.h
@@ -74,7 +74,7 @@ typedef struct {
 
 
 BTL_EVENT_FACTOR* BTL_EVENT_AddFactor( BtlEventFactorType factorType, u16 subID,
-  BtlEventPriority mainPri, u16 subPri, u8 dependID,
+  BtlEventPriority mainPri, u32 subPri, u8 dependID,
   const BtlEventHandlerTable* handlerTable, u16 numHandlers );
 extern BTL_EVENT_FACTOR* BTL_EVENT_SeekFactor( BtlEventFactorType factorType, u8 pokeID );
 extern BTL_EVENT_FACTOR* BTL_EVENT_GetNextFactor( BTL_EVENT_FACTOR* factor );
